+++ Pandora/Paradigm/Structure/Graph.hs
@@ -1,15 +1,18 @@
-module Pandora.Paradigm.Structure.Graph (Graph, loose) where
+module Pandora.Paradigm.Structure.Graph (Graph) where
 
 import Pandora.Core.Functor (type (:.:), type (><))
-import Pandora.Core.Morphism ((.))
+import Pandora.Core.Morphism ((.), ($))
 import Pandora.Paradigm.Basis.Edges (Edges (Empty, Overlay))
 import Pandora.Paradigm.Basis.Twister (Twister ((:<)))
-import Pandora.Paradigm.Inventory.Stateful (fold)
-import Pandora.Pattern.Functor.Traversable (Traversable)
+import Pandora.Pattern.Junction.Composition (Composition (Primary, unwrap))
+import Pandora.Pattern.Functor.Covariant (Covariant ((<$>), (<$$>)))
 
 -- | Acyclic graph structure without loops
-type Graph a = Edges :.: Twister Edges >< a
+newtype Graph a = Graph (Edges :.: Twister Edges >< a)
 
--- | Transform any traversable structure into all loose edges graph
-loose :: Traversable t => t a -> Graph a
-loose = fold Empty (\x -> Overlay . (:<) x)
+instance Covariant Graph where
+	f <$> Graph stack = Graph $ f <$$> stack
+
+instance Composition Graph where
+	type Primary Graph a = Edges :.: Twister Edges >< a
+	unwrap (Graph stack) = stack