+++ CHANGELOG.md
@@ -105,3 +105,4 @@
 * Define `Pointable`, `Applicative`, `Bindable` and `Representable` instances for `->`
 * Define infix `Adjoint` operators - `-|` and `|-`
 * Define `Adjoint` instance for `Stateful` and `Storage` datatypes
+* Change `Graph` definition: from type synonymous to newtype